seperate viz into seperate component
